chore: fix tests, consolidate CTI tests
